adding pbt sac test

+++ qdax/baselines/sac_pbt.py
@@ -396,7 +396,7 @@ class PBTSAC(SAC):
         env: Env,
         num_iterations: int,
         env_batch_size: int,
-        grad_updates_per_step: int,
+        grad_updates_per_step: float,
     ) -> Callable:
         """
         Returns the function to update the population of agents.
